backend/requirements.txt
agno==2.5.14
openai==2.30.0
fastapi==0.135.3
uvicorn==0.42.0
python-dotenv==1.2.2
sqlalchemy>=2.0.0
ag-ui-protocol>=0.0.1
pyyaml>=6.0
httpx>=0.27.0
celery[redis]>=5.3.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
mercadopago>=2.2.0
gspread>=6.0.0
